# apps/backend/docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/auth/otp/request": {
      "post": {
        "operationId": "AuthController_requestOtp_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpRequestResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Requested again before the cooldown elapsed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Request an OTP for login or password reset",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/otp/verify": {
      "post": {
        "operationId": "AuthController_verifyOtp_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/OtpVerifyTokensResponseDto"
                    },
                    {
                      "$ref": "#/components/schemas/OtpVerifyResetResponseDto"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "No active OTP, incorrect OTP, or too many attempts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Verify an OTP — logs in (creating the account if new) for purpose=LOGIN, or returns a short-lived reset token for purpose=PASSWORD_RESET",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginWithPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials or inactive account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Login with email/phone + password",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPairDto"
                }
              }
            }
          },
          "401": {
            "description": "Refresh token is invalid, expired, or already revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Exchange a refresh token for a new token pair",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Revoke a refresh token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/password": {
      "put": {
        "operationId": "AuthController_setPassword_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "Not logged in, or currentPassword is incorrect",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Set a password for the first time, or change it (currentPassword required if one already exists)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/password/reset": {
      "post": {
        "operationId": "AuthController_resetPassword_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "Reset token is invalid or expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Reset password using the token from otp/verify (PASSWORD_RESET)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/social/google": {
      "post": {
        "operationId": "AuthController_loginWithGoogle_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SocialLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          },
          "501": {
            "description": "Google sign-in is not yet configured on this environment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Login/register with a Google ID token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/social/apple": {
      "post": {
        "operationId": "AuthController_loginWithApple_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SocialLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          },
          "501": {
            "description": "Apple sign-in is not yet configured on this environment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Login/register with an Apple ID token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "AuthController_me_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticatedUserDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Current authenticated identity",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "SevaMitra API",
    "description": "SevaMitra backend API — see docs/ARCHITECTURE.md",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "TokenPairDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI1ZjJjOWUzNC0yYjdiLTRkM2EtOWYxYS02YTJmMGM4YjkxZDQiLCJwaG9uZU51bWJlciI6Iis5MTk4NzY1NDMyMTAiLCJlbWFpbCI6bnVsbH0.dGVzdC1zaWduYXR1cmUtZXhhbXBsZQ"
          },
          "refreshToken": {
            "type": "string",
            "example": "9c6e2a1f0b7d4e3c8a5f1b2d6e9c0a3f7b4d8e1c2a5f9b0d3e6c8a1f4b7d0e2c9a5f1b3d6e8c0a2f"
          },
          "expiresIn": {
            "type": "number",
            "description": "Access token lifetime, in seconds",
            "example": 900
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "expiresIn"
        ]
      },
      "PublicUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "5f2c9e34-2b7b-4d3a-9f1a-6a2f0c8b91d4"
          },
          "phoneNumber": {
            "type": "object",
            "example": "+919876543210",
            "nullable": true
          },
          "email": {
            "type": "object",
            "example": null,
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "ACTIVE",
              "SUSPENDED",
              "DELETED"
            ],
            "example": "ACTIVE"
          }
        },
        "required": [
          "id",
          "phoneNumber",
          "email",
          "status"
        ]
      },
      "OtpVerifyTokensResponseDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "tokens"
            ],
            "example": "tokens"
          },
          "tokens": {
            "$ref": "#/components/schemas/TokenPairDto"
          },
          "user": {
            "$ref": "#/components/schemas/PublicUserDto"
          }
        },
        "required": [
          "kind",
          "tokens",
          "user"
        ]
      },
      "OtpVerifyResetResponseDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "resetToken"
            ],
            "example": "resetToken"
          },
          "resetToken": {
            "type": "string",
            "description": "Pass this to POST /auth/password/reset",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI1ZjJjOWUzNC0yYjdiLTRkM2EtOWYxYS02YTJmMGM4YjkxZDQiLCJwdXJwb3NlIjoicGFzc3dvcmRfcmVzZXQifQ.dGVzdC1zaWduYXR1cmUtZXhhbXBsZQ"
          },
          "expiresIn": {
            "type": "string",
            "example": "10m"
          }
        },
        "required": [
          "kind",
          "resetToken",
          "expiresIn"
        ]
      },
      "RequestOtpDto": {
        "type": "object",
        "properties": {
          "phoneNumber": {
            "type": "string",
            "example": "+919876543210"
          },
          "purpose": {
            "type": "string",
            "enum": [
              "LOGIN",
              "PASSWORD_RESET"
            ],
            "default": "LOGIN"
          }
        },
        "required": [
          "phoneNumber",
          "purpose"
        ]
      },
      "OtpRequestResponseDto": {
        "type": "object",
        "properties": {
          "expiresInSeconds": {
            "type": "number",
            "description": "How long the OTP stays valid for",
            "example": 300
          }
        },
        "required": [
          "expiresInSeconds"
        ]
      },
      "ErrorDetailDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "BAD_REQUEST"
          },
          "message": {
            "type": "string",
            "example": "Incorrect OTP"
          },
          "details": {
            "type": "object"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetailDto"
          }
        },
        "required": [
          "error"
        ]
      },
      "VerifyOtpDto": {
        "type": "object",
        "properties": {
          "phoneNumber": {
            "type": "string",
            "example": "+919876543210"
          },
          "otp": {
            "type": "string",
            "example": "123456"
          },
          "purpose": {
            "type": "string",
            "enum": [
              "LOGIN",
              "PASSWORD_RESET"
            ],
            "default": "LOGIN"
          }
        },
        "required": [
          "phoneNumber",
          "otp",
          "purpose"
        ]
      },
      "LoginWithPasswordDto": {
        "type": "object",
        "properties": {
          "identifier": {
            "type": "string",
            "example": "admin@sevamitra.in",
            "description": "Email or E.164 phone number"
          },
          "password": {
            "type": "string",
            "example": "correct-horse-battery-staple"
          }
        },
        "required": [
          "identifier",
          "password"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "tokens": {
            "$ref": "#/components/schemas/TokenPairDto"
          },
          "user": {
            "$ref": "#/components/schemas/PublicUserDto"
          }
        },
        "required": [
          "tokens",
          "user"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "example": "9c6e2a1f0b7d4e3c8a5f1b2d6e9c0a3f7b4d8e1c2a5f9b0d3e6c8a1f4b7d0e2c9a5f1b3d6e8c0a2f"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "SetPasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "description": "Required only if the account already has a password set (i.e. this is a change, not a first-time set).",
            "example": "correct-horse-battery-staple"
          },
          "newPassword": {
            "type": "string",
            "example": "new-correct-horse-battery"
          }
        },
        "required": [
          "newPassword"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "resetToken": {
            "type": "string",
            "description": "Short-lived token returned by POST /auth/otp/verify for purpose=PASSWORD_RESET",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI1ZjJjOWUzNC0yYjdiLTRkM2EtOWYxYS02YTJmMGM4YjkxZDQiLCJwdXJwb3NlIjoicGFzc3dvcmRfcmVzZXQifQ.dGVzdC1zaWduYXR1cmUtZXhhbXBsZQ"
          },
          "newPassword": {
            "type": "string",
            "example": "new-correct-horse-battery"
          }
        },
        "required": [
          "resetToken",
          "newPassword"
        ]
      },
      "SocialLoginDto": {
        "type": "object",
        "properties": {
          "idToken": {
            "type": "string",
            "description": "ID token issued by the provider SDK on the client",
            "example": "eyJhbGciOiJSUzI1NiIsImtpZCI6IjEyMzQ1In0.eyJpc3MiOiJhY2NvdW50cy5nb29nbGUuY29tIn0.dGVzdC1zaWduYXR1cmUtZXhhbXBsZQ"
          }
        },
        "required": [
          "idToken"
        ]
      },
      "AuthenticatedUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "5f2c9e34-2b7b-4d3a-9f1a-6a2f0c8b91d4"
          },
          "phoneNumber": {
            "type": "object",
            "example": "+919876543210",
            "nullable": true
          },
          "email": {
            "type": "object",
            "example": null,
            "nullable": true
          }
        },
        "required": [
          "id",
          "phoneNumber",
          "email"
        ]
      },
...
    }
  }
}